Login Cookies
Username and ID are both stored as cookies. There is a test button on the login page that shows what's currently kept in the cookies.
Username cookie is 'username'
ID cookie is 'id'

Note: staff1 and user1 both have an id = 11111111. Use BOTH username AND id to get the correct details.