Purchase Page
Finished, but looks plain. All fields are required before being able to submit. Upon submit, a success alert is given, every book in the cart reduces in quantity by 1, cart is cleared, and page moved back to home page.
Do people need to be signed in under a customer account to purchase? If so, I need to put a function in to check that a customer is logged in before navigating to the purchase page, otherwise they get navigated to the login page.