Add/Edit/Delete
Staff can now add a book to the database, delete from the database, and edit books in the database. Error checks are in place so books can't have the same ISBN, and that all required data is given. Just a few small bits n pieces here and there that need to be done, such as adding an option to use credits, requiring login to be able to purchase, etc.
Not going to merge this as I am not confident I will do it correctly, and will wipe all your work again :/