Search Box
Added a search box on search page. Now users can search for a title, author, or ISBN, or search by genres. 
Customer login is required to access purchase page. Will be redirected to login if not logged in as a customer. Haven't handled what to do when Staff is logged in. Most likely just a prompt to tell them to login as a customer without navigating them to the login page.
I'm probably missing some class styling's, such as class = "book", but I'll wait for you to merge your work first instead of looking through all the code. Will be easier when I can visually see the page.
Still need to add option for credits at purchase page.